:recycle: refactor: Atualizacao do sistema

+++ src/main/java/com/datatech/datatechapi/controller/RestricoesController.java
@@ -11,8 +11,10 @@ import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.geometry.Pos;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
+import org.controlsfx.control.Notifications;
 
 import java.io.IOException;
 import java.net.URL;
@@ -21,6 +23,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.ResourceBundle;
 
+import static com.datatech.datatechapi.util.Alertas.*;
+import static com.datatech.datatechapi.util.Detalhes.mostrarDataHora;
+
 public class RestricoesController implements Initializable {
 
     @FXML
@@ -28,6 +33,9 @@ public class RestricoesController implements Initializable {
 
     @FXML
     private Button btn_voltar;
+
+    @FXML
+    private Label lbl_rodape;
 
     @FXML
     private ComboBox<DiaDaSemana> cbx_diadasemana;
@@ -58,44 +66,71 @@ public class RestricoesController implements Initializable {
         cbx_diadasemana.getItems().addAll(DiaDaSemana.values());
         cbx_horariodaaula.getItems().addAll(HorarioDaAula.values());
         preencherTabela();
-
-
+        lbl_rodape.setText(mostrarDataHora());
     }
 
     @FXML
     void excluirResticao(ActionEvent event) {
-        restricaoDao.removerRestricao(restricoesObservable.get(tbv_restricoes.getSelectionModel().getSelectedIndex()));
+        Restricao restTv = (Restricao) tbv_restricoes.getSelectionModel().getSelectedItem();
+        if (restTv == null)
+            emitirAlertaSelecao();
+        else{
+            restricaoDao.removerRestricao(restricoesObservable.get(tbv_restricoes.getSelectionModel().getSelectedIndex()));
+            emitirAlertaRestricaoExcluids();
+        }
+
         preencherTabela();
     }
 
     @FXML
     void salvarRestricao(ActionEvent event) {
-        Restricao restricao = new Restricao();
-        restricao.setDiaDaSemana(cbx_diadasemana.getValue());
-        restricao.setHorarioDaAula(cbx_horariodaaula.getValue());
-        restricao.setProfessorEmail(LoginController.EMAIL);
-        restricaoDao.cadastrarRestricao(restricao);
-        preencherTabela();
-        limparCampos();
+
+        if (cbx_horariodaaula.getSelectionModel().isEmpty() || cbx_diadasemana.getSelectionModel().isEmpty()) {
+            emitirAlertaRestricaoCamposVazios();
+        } else if (verificarRestricoes()) {
+            emitirAlertaRetricoesJaExiste();
+        } else {
+            Restricao restricao = new Restricao();
+            restricao.setDiaDaSemana(cbx_diadasemana.getValue());
+            restricao.setHorarioDaAula(cbx_horariodaaula.getValue());
+            restricao.setProfessorEmail(LoginController.EMAIL);
+            restricaoDao.cadastrarRestricao(restricao);
+            preencherTabela();
+            limparCampos();
+            emitirAlertaRestricaoSalva();
+        }
+    }
+
+    boolean verificarRestricoes() {
+        restricoes = restricaoDao.buscarRestricao(LoginController.EMAIL);
+        for (Restricao r : restricoes) {
+            if (cbx_diadasemana.getValue().equals(r.getDiaDaSemana()) &&
+                    cbx_horariodaaula.getValue().equals(r.getHorarioDaAula()))
+                return true;
+        }
+        return false;
     }
 
     void visualizarNomeProfessor() {
         Professor professor = new Professor();
         lbl_professor.setText(LoginController.USUARIOLOGADO);
     }
+
     void preencherTabela() {
         tbc_diadasemana.setCellValueFactory(new PropertyValueFactory<>("diaDaSemana"));
         tbc_horariodaaula.setCellValueFactory(new PropertyValueFactory<>("horarioDaAula"));
 
         restricoes = restricaoDao.buscarRestricao(LoginController.EMAIL);
-        restricoesObservable= FXCollections.observableArrayList(restricoes);
+        restricoesObservable = FXCollections.observableArrayList(restricoes);
 
         tbv_restricoes.setItems(restricoesObservable);
     }
-    void limparCampos(){
+
+    void limparCampos() {
         cbx_diadasemana.getSelectionModel().clearSelection();
         cbx_horariodaaula.getSelectionModel().clearSelection();
     }
+
     @FXML
     void voltarMenu(ActionEvent event) throws IOException {
         App.setRoot("views/telamenu.fxml");
